Don't compile to binary

+++ Makefile
@@ -3,12 +3,10 @@ LDFLAGS=-lboost_regex
 O=out
 S=src
 CXX=clang++
-B=test
 
 files=$O/vick-comp-javafx.o   \
 
 all: ${files}
-	${CXX} -o $B $^ ${CFLAGS} ${LDFLAGS}
 
 $O/%.o: $S/%.cc $S/%.hh
 	@mkdir -p $O
@@ -19,8 +17,7 @@ $O/%.o: $S/%.cc
 	${CXX} -o $@ -c $< ${CFLAGS}
 
 clean:
-	rm -R $O
-	rm $B
+	[ ! -d $O ] || rm -R $O
 
 tags:
 	etags `find $S -name '*.cc'`
